requirements.txt
pyserial
Pillow
numpy
psutil
python-socketio
docker
